Removed unnecessary Classfiles. Moved Classfiles to relevant folders Ran Codemaid to clean up code. Added info to ReadMe
