+++ FindaFlick/Menus/Idmenu.cs
@@ -1,24 +1,21 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
 using System.Threading.Tasks;
+using static System.Console;
+
 namespace FindaFlick
 {
-    class Idmenu
+    internal class Idmenu
     {
         public static async Task RunIdMenu()
         {
-            string icon = @" ___                  _         ___   _                   
-(  _`\  _            ( )       (  _`\(_ )  _        ( )    
-| (_(_)(_)  ___     _| |   _ _ | (_(_)| | (_)   ___ | |/') 
-|  _)  | |/' _ `\ /'_` | /'_` )|  _)  | | | | /'___)| , <  
-| |    | || ( ) |( (_| |( (_| || |    | | | |( (___ | |\`\ 
+            string icon = @" ___                  _         ___   _
+(  _`\  _            ( )       (  _`\(_ )  _        ( )
+| (_(_)(_)  ___     _| |   _ _ | (_(_)| | (_)   ___ | |/')
+|  _)  | |/' _ `\ /'_` | /'_` )|  _)  | | | | /'___)| , <
+| |    | || ( ) |( (_| |( (_| || |    | | | |( (___ | |\`\
 (_)    (_)(_) (_)`\__,_)`\__,_)(_)   (___)(_)`\____)(_) (_)
-                                                           
-                                                           
-Here you can search for a movie by ID.
 
+Here you can search for a movie by ID.
 
 ";
 
@@ -28,7 +25,7 @@ Here you can search for a movie by ID.
             switch (HighIndex)
             {
                 case 0:
-                    await MovieIdMenu.RunMenu();
+                    await RunMenu();
                     break;
 
                 case 1:
@@ -38,10 +35,47 @@ Here you can search for a movie by ID.
                 case 2:
                     Environment.Exit(0);
                     break;
+            }
+        }
 
+        public static async Task RunMenu()
+        {
+            string icon = @" ___                  _         ___   _
+(  _`\  _            ( )       (  _`\(_ )  _        ( )
+| (_(_)(_)  ___     _| |   _ _ | (_(_)| | (_)   ___ | |/')
+|  _)  | |/' _ `\ /'_` | /'_` )|  _)  | | | | /'___)| , <
+| |    | || ( ) |( (_| |( (_| || |    | | | |( (___ | |\`\
+(_)    (_)(_) (_)`\__,_)`\__,_)(_)   (___)(_)`\____)(_) (_)
+
+Here you can search for a movie by ID.
+
+";
+            Console.Clear();
+            WriteLine(icon);
+
+            Console.WriteLine("Enter ID number: ");
+            try
+            {
+                int id = Convert.ToInt32(Console.ReadLine());
+                Movie movieResult = await SearchFunc.SearchById(id);
+                if (movieResult != null)
+                {
+                    Clear();
+                    movieResult.showMovieResult();
+                }
+                else
+                {
+                    Console.WriteLine("Sorry, nothing found");
+                }
+            }
+            catch
+            {
+                Console.WriteLine("Not a number...");
             }
 
-
+            WriteLine("Press any key to clear results and return to menu");
+            ReadKey();
+            await Idmenu.RunIdMenu();
         }
     }
 }
