+++ FindaFlick/README.md
@@ -1,0 +1,5 @@
+# FindaFlick
+A console application used to search for information about movies.
+***
+An api key from <https://www.themoviedb.org/documentation/api> is required to use this application.
+![Finda Flickimage](../FindaFlickimage.jpg)
